# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NumVerify Amazon Growth Agency API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List testimonials
      responses:
        "200":
          description: Testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: createTestimonial
      tags: [testimonials]
      summary: Create testimonial (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
  /testimonials/{id}:
    delete:
      operationId: deleteTestimonial
      tags: [testimonials]
      summary: Delete testimonial (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── FAQ ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # ── Auth ──────────────────────────────────────────────────────
...
    Testimonial:
      type: object
      properties:
        id:
          type: integer
        clientName:
          type: string
        clientTitle:
          type: string
        company:
          type: string
        avatarUrl:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        rating:
          type: integer
        text:
          type: string
        metrics:
          type: ["string", "null"]
        featured:
          type: boolean
      required: [id, clientName, clientTitle, company, rating, text, featured]
    TestimonialInput:
      type: object
      properties:
        clientName:
          type: string
        clientTitle:
          type: string
        company:
          type: string
        avatarUrl:
          type: string
        logoUrl:
          type: string
        rating:
          type: integer
        text:
          type: string
        metrics:
          type: string
        featured:
          type: boolean
      required: [clientName, clientTitle, company, rating, text]

    # ── FAQ ───────────────────────────────────────────────────────
...